Working Input
After lots of testing, now there is a working process for inputing data into the Period fields, which are stored in local storage. Next is to implement this throught the app and create a better structure for storing the data so it can be sent to another app. Slowly getting there...